UMA2 Protection API moved to the permission_association module
